requirements.txt
langchain
openai
python-dotenv
chromadb
tiktoken
pypdfium2
unstructured
